pyproject.toml
[tool.poetry]
name = "cookie-website"
version = "0.1.0"
description = ""
authors = ["mshappel <36205722+mshappel@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
psycopg2 = "^2.9.9"
environs = "^11.0.0"
django-heroku = "^0.3.1"
django-bootstrap-datepicker-plus = "^5.0.5"
django-phonenumber-field = "^8.0.0"
phonenumberslite = "^8.13.40"
twilio = "^9.2.3"
django-bootstrap5 = "^24.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
